Refactoring: Separate Forwarder and ForwarderBackgroundService
