+++ src/Krp/DependencyInjection/KubernetesForwarderOptions.cs
@@ -5,6 +5,6 @@ namespace Krp.DependencyInjection;
 
 public class KubernetesForwarderOptions
 {
-    public List<KubernetesEndpoint> Endpoints { get; set; } = new();
-    public List<HttpEndpoint> HttpEndpoints { get; set; } = new();
+    public List<KubernetesEndpoint> Endpoints { get; set; } = [];
+    public List<HttpEndpoint> HttpEndpoints { get; set; } = [];
 }
