+++ src/Krp/Endpoints/HttpProxy/HttpProxyEndpointHandler.cs
@@ -1,5 +1,4 @@
-﻿using Krp.Endpoints;
-using System;
+﻿using System;
 using System.Net;
 using System.Threading.Tasks;
